fix: split voice-surfaces for RSC boundary + gate header search icon
Copilot-round-1 verification caught two issues on the rebased Phase 2
branch (hotfix changes didn't propagate through rebase):

1. voice-surfaces.ts imported chatCompletion at top level, pulling
   prisma/pg into the client bundle (via chat-panel-store → ChatPanel).
   Split into:
   - voice-surfaces.ts — types + DEFAULT_VOICE_SURFACES (client-safe)
   - voice-surfaces.server.ts — generateVoiceSurfaces() (server-only)
   Updated import paths in ai-search route + regenerate-surfaces route +
   test mock.

2. SiteHeaderWrapper.tsx happy-path branch (non-empty labels) passed
   `aiConfigured={aiConfigured}` without the smartSearchEnabled toggle
   gate. Now matches the empty-labels branch: `aiConfigured={smartSearchActive}`.
   Header search icon now correctly hides when toggle is off.
